Refactor DashboardActivity and implement HomeFragment balance logic
Update DashboardActivity to use explicit fragment imports and
standardize the bottom navigation initialization. The activity
now retrieves the username from the intent to provide context
for data loading in child fragments.

Correct the HomeFragment implementation by renaming the class
from MoreFragment and ensuring it inflates the correct layout.
Replace the placeholder export UI logic with a functional
balance summary that calculates the total sum of expenses
retrieved via JsonHelper for the current user.

Modify GoalActivity to cast seek bar progress values to doubles
when creating a Goal object. This ensures data type consistency
with the underlying model before saving the data to storage.

+++ app/src/main/java/com/example/harmonie_budget_app_kt/GoalActivity.kt
@@ -42,8 +42,8 @@ class GoalActivity : AppCompatActivity() {
         })
 
         btnSaveGoal.setOnClickListener {
-            val min = seekMin.progress
-            val max = seekMax.progress
+            val min = seekMin.progress.toDouble()
+            val max = seekMax.progress.toDouble()
             val goal = Goal(min, max)
             JsonHelper.saveGoal(this, username, goal)
             Toast.makeText(this, "Goals saved", Toast.LENGTH_SHORT).show()
